pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Alexandr Obolonkov <itbolonkov2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.6.0"
django = "4.2.11"
django-ckeditor = "6.7.1"
django-js-asset = "2.0.0"
django-taggit = "3.1.0"
django-taggit-serializer = "0.1.7"
djangorestframework = "3.15.1"
djangorestframework-simplejwt = "5.3.1"
pillow = "9.4.0"
pyjwt = "2.6.0"
python-dotenv = "0.21.1"
pytz = "2022.7"
six = "1.16.0"
sqlparse = "0.4.3"
psycopg2 = "^2.9.9"
unidecode = "^1.3.8"
channels = "^4.1.0"
daphne = "^4.1.2"
channels-redis = "^4.2.0"
celery = {extras = ["redis"], version = "^5.4.0"}


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
pre-commit = "^3.6.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 88
multi_line_output = 3
known_first_party = ["config"]